refactor(PhoneVerificationForm): Extract component.
